move Java7 VM plugin to new package
